52Plots/requirements.txt
mysql-connector-python
argparse
black
pandas
matplotlib
PyQt5 
seaborn